Use a boolean to set if api is internal
Use a boolean to set if api is internal.

+++ 4a-hal/4a-hal-controllers/4a-hal-controllers-api-loader.c
@@ -90,7 +90,7 @@ static int HalCtlsInitOneApi(AFB_ApiT apiHandle)
 		return -3;
 
 	// Fill SpecificHalDatadata structure
-	currentCtlHalData->internal = (unsigned int) true;
+	currentCtlHalData->internal = true;
 
 	currentCtlHalData->apiName = (char *) ctrlConfig->api;
 	currentCtlHalData->sndCard = (char *) ctrlConfig->uid;
